Move shared rosters selection to AddParticipants
