+++ client/src/Containers/Create/MakeRooms/AddParticipants.js
@@ -1,21 +1,36 @@
 /* eslint-disable no-console */
 /* eslint-disable react/no-did-update-set-state */
 import React, { useState, Fragment } from 'react';
+import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import SearchResults from 'Containers/Members/SearchResults';
+import { amIAFacilitator } from 'utils';
 import API from 'utils/apiRequests';
-import { Button, InfoBox, Search, Member } from 'Components';
+import { Button, InfoBox, Search, Member, Slider } from 'Components';
 import GenericSearchResults from 'Components/Search/GenericSearchResults';
 import classes from './makeRooms.css';
 
 const AddParticipants = (props) => {
   const { participants, userId, onSubmit, onCancel } = props;
 
+  const userCoursesById = useSelector((state) => state.courses.byId);
+
+  const coursesUserDidNotCreate = Object.values(
+    userCoursesById
+  ).filter((course) => amIAFacilitator(course, userId));
+
+  const coursesByNames = coursesUserDidNotCreate.reduce((acc, curr) => {
+    return { ...acc, [curr.name]: { ...curr } };
+  }, {});
+
   const [initialSearchResults, setInitialSearchResults] = useState([]);
   const [searchResults, setSearchResults] = useState([]);
-
+  const [rosterSarchResults, setRosterSearchResults] = useState(
+    Object.keys(coursesByNames)
+  );
   const [searchText, setSearchText] = useState('');
   const [newParticipants, setNewParticipants] = useState([]);
+  const [isAddingParticipants, setIsAddingParticipants] = useState(true);
 
   const search = (text) => {
     if (text.length > 0) {
@@ -44,6 +59,28 @@ const AddParticipants = (props) => {
     }
   };
 
+  const searchRosters = (text) => {
+    // initial search -- search course names
+    const res = Object.keys(coursesByNames).filter((courseName) =>
+      courseName.includes(text)
+    );
+
+    if (!res.length) {
+      // search facilitators and return an array of course names containing those facilitators
+      coursesUserDidNotCreate.forEach((course) => {
+        if (
+          course.members.some(
+            (mem) =>
+              mem.role === 'facilitator' && mem.user.username.includes(text)
+          )
+        )
+          res.push(course.name);
+      });
+    }
+
+    setRosterSearchResults(res);
+  };
+
   const addParticipant = (_id, username) => {
     // filter out duplicates in the right column (New Participants column)
     if (
@@ -58,6 +95,13 @@ const AddParticipants = (props) => {
 
     setSearchResults((prevState) =>
       prevState.filter((user) => user._id !== _id)
+    );
+  };
+
+  const addParticipantFromRoster = (courseName) => {
+    console.log(`courseName: ${courseName}`);
+    coursesByNames[courseName].members.forEach((mem) =>
+      addParticipant(mem.user._id, mem.user.username)
     );
   };
 
@@ -87,10 +131,10 @@ const AddParticipants = (props) => {
     const prevParticipants = participants.filter(
       (mem) => mem.role === 'participant'
     );
-    const newList = [...prevParticipants, ...newParticipants]
+    const participantsToAdd = [...prevParticipants, ...newParticipants]
       .sort((a, b) => a.user.username.localeCompare(b.user.username))
       .concat(facilitators);
-    onSubmit(newList);
+    onSubmit(participantsToAdd);
     onCancel();
   };
 
@@ -100,26 +144,59 @@ const AddParticipants = (props) => {
         title="Add Participants"
         icon={<i className="fas fa-user-plus" />}
         className={classes.AddParticipants}
+        rightIcons={
+          <Slider
+            action={() => setIsAddingParticipants((prevState) => !prevState)}
+            isOn={!isAddingParticipants}
+            name="addParticipantsSlider"
+            data-testid="addParticipants-slider"
+          />
+        }
+        rightTitle="Shared Roster"
       >
-        <div className={classes.AddParticipants}>
-          <Fragment>
-            <div style={{ fontSize: '12px' }}>
-              <Search
-                data-testid="member-search"
-                _search={search}
-                placeholder="search by username or email"
-              />
-            </div>
-            {searchResults.length > 0 && (
-              <SearchResults
-                searchText={searchText}
-                usersSearched={searchResults}
-                inviteMember={addParticipant}
-                className={classes.AddParticipants}
-              />
-            )}
-          </Fragment>
-        </div>
+        {isAddingParticipants && (
+          <div className={classes.AddParticipants}>
+            <Fragment>
+              <div style={{ fontSize: '12px' }}>
+                <Search
+                  data-testid="member-search"
+                  _search={search}
+                  placeholder="search by username or email"
+                />
+              </div>
+              {searchResults.length > 0 && (
+                <SearchResults
+                  searchText={searchText}
+                  usersSearched={searchResults}
+                  inviteMember={addParticipant}
+                  className={classes.AddParticipants}
+                />
+              )}
+            </Fragment>
+          </div>
+        )}
+
+        {!isAddingParticipants && (
+          <div className={classes.AddParticipants}>
+            <Fragment>
+              <div style={{ fontSize: '12px' }}>
+                <Search
+                  data-testid="roster-search"
+                  _search={searchRosters}
+                  placeholder="search courses to import rosters from"
+                />
+              </div>
+              {Object.keys(coursesByNames).length > 0 && (
+                <GenericSearchResults
+                  itemsSearched={rosterSarchResults}
+                  searchText={searchText}
+                  select={addParticipantFromRoster}
+                  className={classes.AddParticipants}
+                />
+              )}
+            </Fragment>
+          </div>
+        )}
       </InfoBox>
       {newParticipants && (
         <InfoBox title="New Participants" icon={<i className="fas fa-users" />}>
